requirements.txt
asgiref==3.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.3
django-widget-tweaks==1.5.0
idna==3.10
Pillow==11.3.0
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
